move prune command to archiver.prune
