+++ setup.cfg
@@ -123,6 +123,7 @@ per_file_ignores =
     src/borg/archiver/debug.py:F405
     src/borg/archiver/help.py:E501,F405
     src/borg/archiver/keys.py:F405
+    src/borg/archiver/prune.py:F405
     src/borg/archiver/tar.py:F405
     src/borg/cache.py:E127,E128,E402,E501,E722,W504
     src/borg/fuse.py:E402,E501,E722,W504
